Merged PR 1573: Logout from menu
Added menu to the top-right corner of the application header ("Profile" item on the current header in master)

Added 'Logout' menu item to this menu and attached it to the logout method of authentication service.

Menu is styled as in styleguide. 'Logout' text is resolved from the ICE resources.

'Profile' is changed to user name. Authentication service is updated to save all principal data to local storage (to keep user name as well as pid).

Related work items: #7920, #7942