Merged PR 1787: Motor VIN/Plate Lookup and incremental autofill series of make/model/year/engine
This is a combined solution for the lookup of vehicle by VIN and Palate number. The autofill drop downs for vehicle selection are filled with data aggregated into a single JSON and loaded on page load. Then depending waterfall selection dynamic lists are created for the vehicle inputs.
make > model
                       > year
                       > engine

Related work items: #7767, #7768, #8349