Merged PR 2140: Renamed sql queries and their references
Renamed sql queries and their references.
First part of the name is the name of the object that's queried, middle part if it exists clarifies if its a query for client or agent, last part of the name is 'get' or 'get-detailed' to distinguish variants of the query over the same table.

Related work items: #10723